# POC: Tag-team event loop