--- a/common/src/main/java/io/netty/util/concurrent/StandInThread.java
+++ b/common/src/main/java/io/netty/util/concurrent/StandInThread.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2019 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package io.netty.util.concurrent;
+
+import io.netty.util.Recycler;
+
+/**
+ * Not for general use. Implemented by threads which should be treated
+ * as if they are the same as some other owning thread for thread-local
+ * purposes (i.e. where only one of the two can be running arbitrary
+ * code at a time).
+ * <p>
+ * This only exists to allow the {@link Recycler} to work effectively
+ * in thread tag-team scenarios.
+ */
+public interface StandInThread {
+
+    Thread mainThread();
+}