Motivation

This PR is for discussion/feedback and is in a mostly-working but unfinished state. The idea is to use two threads per event loop, one waits on epoll_wait and the other waits using `LockSupport.park/parkNanos`. Whichever wakes up first "becomes" the event loop and processes the task queue. Efficient coordination is achieved via tight integration with a specially customized queue implementation, derived from JCTools' [`MpscBlockingConsumerArrayQueue`](https://github.com/JCTools/JCTools/blob/master/jctools-core/src/main/java/org/jctools/queues/MpscBlockingConsumerArrayQueue.java). IO events can be received and enqueued by the epoll thread even while the other thread is active. Time-scheduled task expiry is handled via the primary/non-epoll thread only.

~Latency advantages are similar to those targeted when using epoll busy-wait (see #8267), but without needing to dedicate an entire core per event loop.~ It's also partially related to the improvements in #7834.

Benefits
- IO round-trip latency, at least for certain cases/workloads - see benchmarks
- Much more responsive task submission from outside the event loop
- Much cheaper time-based task scheduling from outside event loop
- Appears to reduce CPU for equivalent work (e.g. fewer syscalls)
- Simpler native code interactions - timerfd and eventfd not required/used (eventfd used just for shutdown)
- No need to wake up every second

Some other notes
- This was done with edge-triggered in mind, but I *think* it should work for level-triggered too, provided the event array hand-off buffer size is set to 2 (effectively 1). Doing so doesn't appear to affect performance much, maybe it would make a bigger difference under higher load/concurrency.
- It's currently passing all but 2 of 366 unit tests; I think one of the failures is expected/ok, the other is `EpollSpliceTest::spliceToSocket` which passes when run standalone but not as part of the suite.
- The queue is currently bounded/fixed-size and uses `Unsafe`. I'm *fairly sure* other JCTools MPSC queues could be similarly adapted as required (chunked/growable/unbounded, `Unsafe`-free)
- Various other loose ends remain (e.g. error handling in some places)


Modifications

The new queue class has been named `TagTeamConsumerArrayQueue` for now. It is always in one of three distinct states: "primary" thread active, "secondary" (epoll) thread active, neither active. This state is encoded in the last 2 bits of its long `producerIndex` - a kind of generalization of how the blocked-consumer state is encoded in the original `MpscBlockingConsumerArrayQueue`. The public methods fall into one of four categories: safe for any caller, called only from primary consumer thread, called only from secondary consumer thread, called only from currently-active consumer thread.

The only existing class with significant changes is `EpollEventLoop`. epoll-thread specific logic has been grouped into an `EpollLoop` inner class. The main thread deals with opaque tasks only. A circular queue of `EpollEventArray`s is used to hand off received events to the task processing thread. The epoll thread will park until the other thread catches up if this wraps. There are no JNI code changes.


Results

Adjustments were also made to the benchmark code, it turns out some were dominated by measurement overhead.

Before:
```
Benchmark                                   Mode  Cnt        Score        Error  Units
EpollSocketChannelBenchmark.pingPong       thrpt   20    36947.625 ±    246.654  ops/s
EpollSocketChannelBenchmark.executeSingle  thrpt   20   266313.636 ±  12611.497  ops/s
EpollSocketChannelBenchmark.executeMulti   thrpt   20  6746333.696 ± 674330.117  ops/s
```

After:
```
Benchmark                                   Mode  Cnt        Score        Error  Units
EpollSocketChannelBenchmark.pingPong       thrpt   20    42110.442 ±    546.547  ops/s
EpollSocketChannelBenchmark.executeSingle  thrpt   20  4267783.830 ± 179377.697  ops/s
EpollSocketChannelBenchmark.executeMulti   thrpt   20  9609216.911 ± 645376.994  ops/s
```


Before:
```
Benchmark                                   (burstLength)  (work)  Mode  Cnt     Score     Error  Units
BurstCostExecutorsBenchmark.test1Producer               1       0  avgt   10  2749.784 ±  20.207  ns/op
BurstCostExecutorsBenchmark.test1Producer               1      10  avgt   10  2903.637 ± 252.079  ns/op
BurstCostExecutorsBenchmark.test1Producer              10       0  avgt   10  3080.863 ± 476.379  ns/op
BurstCostExecutorsBenchmark.test1Producer              10      10  avgt   10  3809.870 ± 580.936  ns/op
BurstCostExecutorsBenchmark.test2Producers              1       0  avgt   10  2752.020 ± 127.169  ns/op
BurstCostExecutorsBenchmark.test2Producers              1      10  avgt   10  2641.059 ± 111.378  ns/op
BurstCostExecutorsBenchmark.test2Producers             10       0  avgt   10  2963.372 ± 347.863  ns/op
BurstCostExecutorsBenchmark.test2Producers             10      10  avgt   10  1973.417 ±  45.339  ns/op
BurstCostExecutorsBenchmark.test3Producers              1       0  avgt   10  2087.238 ± 257.431  ns/op
BurstCostExecutorsBenchmark.test3Producers              1      10  avgt   10  1821.122 ±  76.298  ns/op
BurstCostExecutorsBenchmark.test3Producers             10       0  avgt   10  2623.262 ±  37.176  ns/op
BurstCostExecutorsBenchmark.test3Producers             10      10  avgt   10  2439.716 ±  42.170  ns/op
```

After:
```
Benchmark                                   (burstLength)  (work)  Mode  Cnt     Score      Error  Units
BurstCostExecutorsBenchmark.test1Producer               1       0  avgt   10  3171.981 ± 1803.802  ns/op
BurstCostExecutorsBenchmark.test1Producer               1      10  avgt   10  1497.771 ± 1547.308  ns/op
BurstCostExecutorsBenchmark.test1Producer              10       0  avgt   10  2952.119 ± 1581.608  ns/op
BurstCostExecutorsBenchmark.test1Producer              10      10  avgt   10  2594.738 ± 1333.943  ns/op
BurstCostExecutorsBenchmark.test2Producers              1       0  avgt   10   215.261 ±    7.909  ns/op
BurstCostExecutorsBenchmark.test2Producers              1      10  avgt   10   227.148 ±    7.375  ns/op
BurstCostExecutorsBenchmark.test2Producers             10       0  avgt   10  1388.988 ±  128.052  ns/op
BurstCostExecutorsBenchmark.test2Producers             10      10  avgt   10   987.340 ±   15.930  ns/op
BurstCostExecutorsBenchmark.test3Producers              1       0  avgt   10   242.427 ±   15.450  ns/op
BurstCostExecutorsBenchmark.test3Producers              1      10  avgt   10   256.689 ±   14.192  ns/op
BurstCostExecutorsBenchmark.test3Producers             10       0  avgt   10  2368.841 ±  106.535  ns/op
BurstCostExecutorsBenchmark.test3Producers             10      10  avgt   10  1811.361 ±  583.094  ns/op
```