@franz1981 2019-06-24
I need some time to digest and understand it properly: I'm a lil surprised about the perf increase, considering that is using 2 threads: `BurstCostExecutorsBenchmark` is showing very weird numbers and almost no difference while using several producers threads...that is VERY weird...