@njhill 2019-06-24
@franz1981 I agree the `BurstCostExecutorsBenchmark` numbers look a bit weird, and have been looking a bit more into what might be going on. I think a key question is what `BurstCostExecutorsBenchmark` is really aiming to measure... at least in the single producer case I had thought "responsiveness from a 'cold-start' to a new task or burst of such tasks", however as currently implemented I don't think that's what the results really reflect, due to the fact that the EL may not have returned to its original state prior to the subsequent burst coming in.

It means for example that "slower" impls could show _much_ better results just due to the fact that they hang around in the event loop long enough to see the next burst before attempting to go to sleep. And this must be why the higher producer-count ones have better numbers than single consumer (EL is kept awake more of the time).

In the results for this PR above, my current suspicion is that the slowness is primarily due to contention between concurrent producer and consumer CASes of the producerIndex ... former for the new task, latter the EL attempting to enter blocking state after the prior task. It's like we need to use [`Iteration.Level`](http://javadox.com/org.openjdk.jmh/jmh-core/1.12/org/openjdk/jmh/annotations/Level.html#Iteration) but the warnings around that are quite scary and warning 4 may mean this doesn't solve the problem anyhow. Maybe better is to spin or burn for a "precise" amount of time between each iteration, which can then be subtracted?