@franz1981 2019-06-25
> however as currently implemented I don't think that's what the results really reflect, due to the fact that the EL may not have returned to its original state prior to the subsequent burst coming in.

Totally agree: I suppose at this point that the benchmark needs to be redesigned to account of it...or maybe not :P
The whole point of this benchmark was to measure/observe the "user experience" and that's what a user would experience indeed; if that's obtained by pure luck (of timing) that's an implementation detail of the EL. I think that the benchmark need to provide a way to linger right after the burst execution in order to observe how different lingering would affect it (by using BlackHole::consumerCPU after the method execution) and user JMH `Control` to mark the end of the measurement. Let me open a PR for this :+1: 
While re the multi threaded case I believe that the issue is different and was explained on https://github.com/netty/netty/pull/9078: the multi producer case is not correctly awaiting the whole burst to be executed