@johnou 2019-06-25
@njhill if the thread changes for the event loop doesn't that mean external sync is required for managing the client session state?