@njhill 2019-06-25
@johnou it shouldn't since only one performs event loop activity at a time and there are appropriate memory barriers between one going to sleep and the other waking up. The same ThreadLocals are also shared between the two. Most code should not know the difference. That said there could be some which relies on custom Thread impls from custom ThreadFactory/Executors which would break, and not all the unit tests are passing yet so it's possible there's something else I've missed! Within netty, the `Recycler` does actually use the current thread's identity, but this just means it might not recycle as effectively and I have a small tweak which rectifies that too (omitted here for simplicity).