@njhill 2019-06-25
Pushed a few small updates including aforementioned `Recycler` support and a fix for a bug which was causing some of the unit test failures and I found that the multicast tests fail on my system regardless. So there are now only two failures, both intermittent. One is a race around setting auto-read from outside the EL, which I think is probably expected/ok. The other is `EpollSpliceTest::spliceToSocket` timing-out, which I suspect might be LT-related.