@tjake 2019-06-26
My thought looking at the patch is this might improve latency for "light" workloads. but under load the fact that both threads can be working at the same time would make it perform worse.  I think Netty allocates 2x the cores for the default EventLoopGroup.  Perhaps cutting this in half for this kind of approach would be needed.  In anycase is it possible to add a microbench for saturation?