@franz1981 2019-06-26
My concern instead is about the type of queue used: this mpsc variant is using CAS on producer index to communicate/manage a blocked status; that prevent using other types of queues (eg the XADD mpsc I've recently added on jctools) more suited to cover efficiently other type of loads ie multiple producers.

Under heavy load it is very unlikely that the event loop is going idle (hopefully) and so the benefit of this implementation tends to decrease, as noted by @tjake.
But I'm curious to see the under load perf results...
Anyway you've done a great work here Nick, as always :)