@njhill 2019-06-26
Thanks @tjake @franz1981.

The only "work" which which may occur on the second thread is waking up from epoll_wait and adding a task to the queue. I can see that this could result in a higher frequency of wakeups each containing a smaller number of events, but expect if max tput is impacted then it could be mitigated/solved by just setting the buffer of event arrays to be small (e.g. 2). There could easily be a gap in my understanding, but if we expect the ELs to really be 100% busy, why is the default to have 2x the number of cores? I do agree it would be nice to have some saturated tput tests though.

Right now the characteristics of a "fully busy" EL will likely be different depending on which thread has the reins, i.e. in one case the "inactive" thread is feeding IO events to the first, in the other case it's doing nothing. If one of these cases turns out to be superior then I think it shouldn't be too hard to modify things to make the roles of the two threads fluid, i.e. for optimal steady state whichever happened to wake up first.

@franz1981 I think the CAS is kind of required in our use case to guard against unnecessary unparks, the cost of which I assume would outweigh perf benefit of XADD over CAS. That said I think there may be opportunities to incorporate XADD in some places... i.e. adapt your queue (which would involve adding some CAS) or add some XADD into this one. It would be a challenge to do that for the bounded case though as you know. I wonder if anyone actually uses a bounded task queue.