@belliottsmith 2019-07-01
> That's actually not the case, this patch does already use parkNanos to do the timing (totally understandable that you missed this)...

Except that, as mentioned, the current behaviour is a bug - strictly worse than if the eventLoop has a single scheduled task.  So it's definitely a confounder to measure against a known bug.

>  if the consumer is not yet waiting, and the consumer side does not leave user space if there's an un-consumed wakeup pending

This sounds a lot like the race I was originally discussing?  If the consumer is not yet waiting, it is definitely _about_ to wait, so there is a strong chance of competing for the lock that guards the thread's runnable status?

Any cost accounting here has to look at the context, and in this case these can happen strictly no more often than legitimate suspend/awake cycles, so we're not able to ever cost more than 100% as a function of the thing we're trying to improve.  But since the race window is very narrow, it is likely a very small percentage, and perhaps not even measurable, even discounting the above increased cost in this race.  

I think we would need to have a strong indication that this is a real problem before redesigning so significantly.

> I don't quite follow this... the benchmark-stated variance is actually slightly lower as a proportion of the mean (4.2% compared to 4.7%). Did you mis-count the digits or have I misunderstood what you mean?

From your earlier post:

Before

```
BurstCostExecutorsBenchmark.test1Producer               1       0  avgt   10  2749.784 ±  20.207  ns/op
BurstCostExecutorsBenchmark.test1Producer               1      10  avgt   10  2903.637 ± 252.079  ns/op
BurstCostExecutorsBenchmark.test1Producer              10       0  avgt   10  3080.863 ± 476.379  ns/op
BurstCostExecutorsBenchmark.test1Producer              10      10  avgt   10  3809.870 ± 580.936  ns/op
```

After
```
BurstCostExecutorsBenchmark.test1Producer               1       0  avgt   10  3171.981 ± 1803.802  ns/op
BurstCostExecutorsBenchmark.test1Producer               1      10  avgt   10  1497.771 ± 1547.308  ns/op
BurstCostExecutorsBenchmark.test1Producer              10       0  avgt   10  2952.119 ± 1581.608  ns/op
BurstCostExecutorsBenchmark.test1Producer              10      10  avgt   10  2594.738 ± 1333.943  ns/op
```

This looks like a 50% measurement error after, and 10-20% error before.  What am I missing?


> but my suggestion is that the efficiency savings are helping to allow this to occur, i.e. it would not be possible to the same degree without the futex properties (still hypothesis though of course...).

But this can only happen in rare race scenarios, or at least not _ordinarily_.  So how is it happening _reliably_ - with measurement error as low as ~10ns?  This suggests something that is happening with much greater reliability than a race of this kind, so it seems a poor explanation for the behaviour.